Created moduleInfo + added template for Drag and Drop
Template_DragDropFileExample.py - file just dumped in here for now since
it uses signals to identify the file that has been dropped, which is
really handy for dropping in the module information directly from a
.xlsx file

moduleInfo has also been added but not tested. This is currently
building a moduleData class that deals with stripping the info out of a
Sugden module .xlsx file. It will take the main module data. Find the
student data and strip that out too... Needs testing on every level
since it is just blocked in for now.