moduleInfo fed into myWidgets and custom TreeWidget
Custom TreeWidget now uses the moduleInfo.moduleData to strip out the
student list from the template file. This data strip will be upgraded to
occur from a drag and dropped file, rather than this template, that is
in the root of the package (SFX5000_Report.xlsx). The headers for the TW
are set in the class, but need to add functionality to add and remove
some of the extra columns