Order of columns maintained using OrderedDict
from collections import OrderedDict

This neat trck allows you to create a dictionary from a list of pairs as
an option in the constructor. This will build a dictionary, but one that
will remembers the order of the creation, so it can be cycled through.
This means that we can maintain our column order when ever we need to...

TreeWidget width has been expanded, but this will be updated, so that
the width calculates off the sum of the column widths, depending on what
is visible...